Refactor Signup process
